Move FeeBreakdown widgets into their own file
